Move nullptr check to beginning of QInputDevice::operator<<
Nullptr check was performed after aquisition of the d-pointer.
That acquisition crashes if nullptr is passed to the operator, so the
actual check was never hit.

This patch moves the nullptr check to the beginning of the method.

Fixes: QTBUG-112174
Change-Id: If339e2de9ce2e33e10d925e79ca06b3854a24f76
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit f5c55da04bc53bcaf93bdb34588ef21af10333c6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>