Remove hellovulkanwindow example
The idea being that hellovulkantriangle demonstrates the same things.
As a getting started tutorial hellovulkanwindow is the ideal example,
but then again QVulkanWindow is not something we want to promote much
in Qt 6.

Some of the docs are moved to hellovulkantriangle.

Change-Id: Icbfff70b4a4c7e4c0863a937f3c16038c0b03fbe
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 0e8d347c6a99c2729eb5d32cfc70f4fb2755f940)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>